requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-stubs==5.0.0
django-stubs-ext==5.0.0
django-types==0.19.1
djangorestframework==3.15.1
djangorestframework-types==0.8.0
idna==3.7
psycopg2-binary==2.9.9
requests==2.31.0
sqlparse==0.5.0
types-psycopg2==2.9.21.20240417
types-PyYAML==6.0.12.20240311
types-requests==2.31.0.20240406
typing_extensions==4.11.0
urllib3==2.2.1
